+++ examples/configs/src/view-configs/kpmp-premiere.js
@@ -329,7 +329,7 @@ function PageComponent() {
           </div>
           <div className={clsx('view-row', 'view-row-tall')}>
             <div className="view-row-left">
-              <p>This view displays differential expression test results. The arrows on the bottom left and bottom right denote the direction of the effect. Click a point in the plot to select the corresponding gene. <br/><br/>Note that differential expression tests have been run for each cell type separately, so the each gene can appear multiple times (once per cell type).</p>
+              <p>This view displays differential expression test results. The arrows on the bottom left and bottom right denote the direction of the effect. Click a point in the plot to select the corresponding gene. <br /><br />Note that differential expression tests have been run for each cell type separately, so the each gene can appear multiple times (once per cell type).</p>
             </div>
             <div className="view-row-center">
               <VolcanoPlot />
