+++ packages/view-types/statistical-plots/src/VolcanoPlot.js
@@ -5,108 +5,9 @@ import { scaleLinear } from 'd3-scale';
 import { axisBottom, axisLeft } from 'd3-axis';
 import { extent as d3_extent } from 'd3-array';
 import { select } from 'd3-selection';
-import { isEqual } from 'lodash-es';
 import { capitalize, getDefaultForegroundColor } from '@vitessce/utils';
 import { colorArrayToString } from '@vitessce/sets-utils';
-import { getColorScale } from './utils.js';
-
-// Data transformation hook function that is used both here
-// and in the FeatureStatsTable view.
-export function useFilteredVolcanoData(props) {
-  const {
-    data,
-    obsSetsColumnNameMapping,
-    sampleSetsColumnNameMapping,
-    featurePointFoldChangeThreshold,
-    featurePointSignificanceThreshold,
-    sampleSetSelection,
-  } = props;
-
-  const obsSetsColumnNameMappingReversed = useMemo(() => Object.fromEntries(
-    Object
-      .entries(obsSetsColumnNameMapping)
-      .map(([key, value]) => ([value, key])),
-  ), [obsSetsColumnNameMapping]);
-
-  const sampleSetsColumnNameMappingReversed = useMemo(() => Object.fromEntries(
-    Object
-      .entries(sampleSetsColumnNameMapping)
-      .map(([key, value]) => ([value, key])),
-  ), [sampleSetsColumnNameMapping]);
-
-  const computedData = useMemo(() => {
-    return data.map(d => {
-      const { metadata } = d;
-
-      const coordinationValues = metadata.coordination_values;
-
-      const rawObsSetPath = coordinationValues.obsSetFilter
-        ? coordinationValues.obsSetFilter[0]
-        : coordinationValues.obsSetSelection[0];
-      const obsSetPath = [...rawObsSetPath];
-      obsSetPath[0] = obsSetsColumnNameMappingReversed[rawObsSetPath[0]];
-
-      // Swap the foldchange direction if backwards with
-      // respect to the current sampleSetSelection pair.
-      // TODO: move this swapping into the computedData useMemo?
-      let shouldSwapFoldChangeDirection = false;
-      if (
-        coordinationValues.sampleSetFilter
-        && coordinationValues.sampleSetFilter.length === 2
-      ) {
-        const rawSampleSetPathA = coordinationValues.sampleSetFilter[0];
-        const sampleSetPathA = [...rawSampleSetPathA];
-        sampleSetPathA[0] = sampleSetsColumnNameMappingReversed[rawSampleSetPathA[0]];
-
-        const rawSampleSetPathB = coordinationValues.sampleSetFilter[1];
-        const sampleSetPathB = [...rawSampleSetPathB];
-        sampleSetPathB[0] = sampleSetsColumnNameMappingReversed[rawSampleSetPathB[0]];
-
-        if (
-          isEqual(sampleSetPathA, sampleSetSelection[1])
-          && isEqual(sampleSetPathB, sampleSetSelection[0])
-        ) {
-          shouldSwapFoldChangeDirection = true;
-        }
-      }
-
-      return ({
-        ...d,
-        df: {
-          ...d.df,
-          minusLog10p: d.df.featureSignificance.map(v => -Math.log10(v)),
-          logFoldChange: d.df.featureFoldChange.map(v => Math.log2(v) * (shouldSwapFoldChangeDirection ? -1 : 1)),
-        },
-      });
-    });
-  }, [
-    data, obsSetsColumnNameMappingReversed, sampleSetsColumnNameMappingReversed,
-    sampleSetSelection,
-  ]);
-
-  const filteredData = useMemo(() => computedData.map((d) => ({
-    ...d,
-    // Instead of an object of one array per column,
-    // this is now an array of one object per row.
-    df: d.df.featureId.map((featureId, i) => ({
-      featureId,
-      logFoldChange: d.df.logFoldChange[i],
-      featureSignificance: d.df.featureSignificance[i],
-      minusLog10p: d.df.minusLog10p[i],
-      // TODO: include the obsSetName.
-    })).filter(d => (
-      (Math.abs(d.logFoldChange) >= (featurePointFoldChangeThreshold ?? 1.0))
-      && (d.featureSignificance <= (featurePointSignificanceThreshold ?? 0.05))
-    ))
-  })), [computedData, featurePointFoldChangeThreshold, featurePointSignificanceThreshold]);
-
-  return [
-    computedData,
-    filteredData,
-    obsSetsColumnNameMappingReversed,
-    sampleSetsColumnNameMappingReversed
-  ];
-}
+import { getColorScale, useFilteredVolcanoData } from './utils.js';
 
 export default function VolcanoPlot(props) {
   const {
